requirements.txt
requests
websockets
PyYAML
pandas